Begin test framework
The general idea is to write tests in shell, giving us an easy
and powerful system for setting up file- and directory-based tests
(which are the domains in which 11pm operates). C can be called from
inside the shell script.

The general test utilites are contained in a program called `t`.
This program will be implemented in `src/test/t.c`, and complied
along with a `test/*/t_calls.c` file.

The `t_calls.c` file is generated from all the test scripts by
`src/test/make_t_calls.sh`, and provides the inline C so the `t`
utility may make use of it. The syntax for inline C is thus:

  ## include "foo.h"
  #= some_single_c_call_to_check_result();
  t eval @@ : assert -some-assert-on-return

  #{
  #   struct some_type_t some_var;
  #   some_setup_function(&some_var);
  #   return do_some_operation(&some_var);
  #}
  t eval @@ : assert -some-assert-on-return